refactor(runtime): unify MakaTool imports on tool-runtime.js single source
MakaTool/MakaToolContext are defined in tool-runtime.ts; ai-sdk-backend.ts
only re-exported them. Repoint every consumer (session-manager,
runtime-kernel, builtin-tools, and the ai-sdk-backend/session-manager/
deferred-tools tests) to import the types directly from tool-runtime.js,
export them from the runtime barrel via tool-runtime.js, and drop the
redundant re-export from ai-sdk-backend.ts so there is one canonical source.
Type-only change; no runtime behavior change.
