+++ packages/runtime/src/session-manager.ts
@@ -63,7 +63,7 @@ import {
 } from './terminal-run-commit.js';
 
 import type { AgentBackend } from '@maka/core/backend-types';
-import type { MakaTool } from './ai-sdk-backend.js';
+import type { MakaTool } from './tool-runtime.js';
 import type { RunTraceRecorder } from './run-trace.js';
 import type { ShellRunProcessManager } from './shell-run-manager.js';
 import type { ActiveFullCompactBlock } from './active-full-compact.js';
